# System Integrator/system-openapi-v1.0.yaml
openapi: "3.0.3"
info:
  title: 网络小说智能推荐系统 API
  description: |
    集趋势洞察、语义搜索、舆情分析为一体的网络小说智能推荐系统全局 API。
    本规范定义 System Integrator 对外暴露的统一接口及各子模块的内部接口契约。
  version: "1.0.0"
  contact:
    name: System Integrator - 苗文昊
servers:
  - url: http://localhost:8000/api/v1
    description: 集成网关 (开发环境)
  - url: http://localhost:8001/api/v1
    description: Trend Explorer Mock (吴隐)
  - url: http://localhost:8002/api/v1
    description: Safe-Search Architect Mock (吴宸)
  - url: http://localhost:8003/api/v1
    description: Sentiment Critic Mock (武文杰)
...
paths:
...
  /search/semantic:
    post:
      tags: [Search]
      summary: 语义搜索
      operationId: semanticSearch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchRequest"
      responses:
        "200":
          description: 搜索结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResult"
...
  /sentiment/detail/{uid}:
    get:
      tags: [Sentiment]
      summary: 单书舆情详情
      operationId: getSentimentDetail
      parameters:
        - name: uid
          in: path
          required: true
          schema:
            type: string
            pattern: "^[a-f0-9]{64}$"
      responses:
        "200":
          description: 舆情详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SentimentData"
        "404":
          description: UID 未找到
  /sentiment/compare:
    get:
      tags: [Sentiment]
      summary: 多书舆情对比
      operationId: compareSentiment
      parameters:
        - name: uids
          in: query
          required: true
          schema:
            type: string
            description: 逗号分隔的 UID 列表
          example: "abc123...,def456...,ghi789..."
      responses:
        "200":
          description: 对比结果
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SentimentData"

  # ================================================================
  # 运维接口
  # ================================================================
...
components:
  schemas:
...
    SearchRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
          description: 自然语言搜索语句
        filters:
          type: object
          properties:
            tags_include:
              type: array
              items:
                type: string
            tags_exclude:
              type: array
              items:
                type: string
            min_heat:
              type: number
            platforms:
              type: array
              items:
                type: string
        safe_tags:
          type: array
          items:
            type: string
          description: 避雷标签列表
        top_k:
          type: integer
          default: 10
    SearchResult:
      type: object
      properties:
        results:
          type: array
          items:
            type: object
            properties:
              uid:
                type: string
              metadata:
                $ref: "#/components/schemas/Metadata"
              relevance_score:
                type: number
              match_reason:
                type: string
              safe_check:
                type: object
                properties:
                  passed:
                    type: boolean
                  warnings:
                    type: array
                    items:
                      type: string
        query_rewrite:
          type: string
        total_hits:
          type: integer
    SentimentData:
      type: object
      properties:
        uid:
          type: string
        metadata:
          $ref: "#/components/schemas/Metadata"
        sentiment_scores:
          $ref: "#/components/schemas/SentimentScores"
        critic_summary:
          $ref: "#/components/schemas/CriticSummary"
        review_stats:
          $ref: "#/components/schemas/ReviewStats"
    Metadata:
      type: object
      properties:
        title:
          type: string
        platform:
          type: string
        last_update:
          type: string
          format: date-time
    SentimentScores:
      type: object
      properties:
        overall:
          type: number
          minimum: 0
          maximum: 10
        style:
          type: number
        logic:
          type: number
        character:
          type: number
        update_stability:
          type: number
        toxicity_index:
          type: number
          minimum: 0
          maximum: 1
    CriticSummary:
      type: object
      properties:
        one_liner:
          type: string
        pros:
          type: array
          items:
            type: string
        cons:
          type: array
          items:
            type: string
    ReviewStats:
      type: object
      properties:
        total_count:
          type: integer
        positive_ratio:
          type: number
        negative_ratio:
          type: number
        source_breakdown:
          type: object